tests

+++ tests/test_tree.py
@@ -18,9 +18,9 @@ def test_tree(root_name: Any) -> None:
 
     A.create_array(name="x", shape=(2), dtype="float64")
     A.create_array(name="y", shape=(0,), dtype="int8")
-    B.create_array(name="x", shape=(0,))
-    C.create_array(name="x", shape=(0,))
-    D.create_array(name="x", shape=(0,))
+    B.create_array(name="x", shape=(0,), dtype="float64")
+    C.create_array(name="x", shape=(0,), dtype="float64")
+    D.create_array(name="x", shape=(0,), dtype="float64")
 
     result = repr(g.tree())
     root = root_name or ""
